+++ src/index.jsx
@@ -1,10 +1,9 @@
 var api = require('./index.js');
 api.DefaultLoader = require('./DefaultLoader');
-api.loader.addLoader(require('./ValidatorLoader'));
-api.loader.addLoader(require('./DefaultLoader'));
+api.decorators.provide.defaultLoader = api.loader;
+api.loader.addLoader(api.DefaultLoader);
 api.templates = require('./templates/index.js');
 api.types = require('./types/index.js');
 api.styles = require('./styles/index.js');
 api.Dom = require('./Dom');
-api.decorators = require('./decorators');
 module.exports = api;
