+++ src/types/Hidden.jsx
@@ -4,6 +4,7 @@ import {field} from '../decorators';
 @field
 export default class Hidden extends Component {
     render() {
-        return <input id={this.props.name} name={this.props.name} type="hidden" value={JSON.stringify(this.state.value)}  />
+        return <input id={this.props.name} name={this.props.name} type="hidden"
+                      value={JSON.stringify(this.state.value)}/>
     }
 }
